fixes variant process type implemented bwa mapper configured variant quality recalibration drafted
