+++ rules/mapping/mapper/bwa.snake
@@ -18,11 +18,12 @@ rule bwa__map_reads_to_reference:
     output:
         bam   = configured_temp('mapping/{reference}/original/{sample}.bam')
     log:
-        err   = 'mapping/{reference}/original/log/{sample}.bwa.map_stats'
+        out   = 'mapping/{reference}/original/log/{sample}.bam.out',
+        err   = 'mapping/{reference}/original/log/{sample}.bam.err'
     params:
         index = 'reference/{reference}/bwa_index/{reference}',
         additional = method_config.get('params', ''),
-        concordant = '-B 40 -O 60 -E 10  –L100' if method_config.get('only_concordant', False) else ''
+        # concordant = '-B 40 -O 60 -E 10  –L100' if method_config.get('only_concordant', False) else ''
     threads:
         int(config['threads'])
     conda:
@@ -34,9 +35,7 @@ rule bwa__map_reads_to_reference:
             {params.index} \
             {input.r1} \
             {input.r2} \
-            {params.concordant} \
-            {params.additional} \
-        2> {log.err} \
-        | samtools view -bS - \
-        > {output.bam}
+            -M \
+            2> {log.err} \
+        | samtools view -bS > {output.bam} \
         """
