+++ rules/reference/annotation/regions.snake
@@ -46,3 +46,5 @@ rule picard__bed_to_interval_list:
             1> {log.out} \
             2> {log.err}
         """
+
+
