+++ rules/variant/report/calling/gatk.snake
@@ -1,32 +1,3 @@
-rule gatk__fix_vcf_header:
-    """
-    Adds missing sequence dictionary to VCF header. This job also generates VCF index (.idx).
-    :input vcf: Raw VCF from variant caller vardict.
-    :input fasta: Reference sequence.
-    :output vcf: Fixed VCF.
-    :output vcf_index: Fixed VCF index.
-    """
-    input:
-        vcf = 'variant/{reference}-{panel}/original/{sample}.vardict.vcf',
-        fasta = 'reference/{reference}/{reference}.fa'
-    output:
-        vcf = 'variant/{reference}-{panel}/original/{sample}.vcf',
-        vcf_index = 'variant/{reference}-{panel}/original/{sample}.vcf.idx',
-    log:
-        out = 'variant/{reference}-{panel}/original/log/{sample}.out',
-        err = 'variant/{reference}-{panel}/original/log/{sample}.err',
-    conda:
-        config['snakelines_dir'] + '/enviroments/others_env.yaml'
-    shell:
-        """
-        gatk SelectVariants \
-            -V {input.vcf} \
-            -R {input.fasta} \
-            -O {output.vcf} \
-            1>{log.out} \
-            2>{log.err}
-        """
-
 rule tabix__index_vcf:
     """
     Create tabix intex on BGZF (bgzipped) VCF file.
