+++ src/common.snake
@@ -1,7 +1,7 @@
 import sys
 import tempfile
 from pipeline.samples import Pipeline
-from config.impute_mapping_types import impute_mapping_types, impute_read_types
+from config.impute_process_types import impute_read_types, impute_mapping_types, impute_variant_types
 from config.load_imports import import_files_in_config
 from config.values_retrieval import extract_from_config
 
@@ -35,8 +35,9 @@ for sample_configuration in sample_configurations:
 assert not pipeline.is_empty(), 'No samples to analyse'
 
 # impute types:
-impute_mapping_types(pipeline, config)
 impute_read_types(pipeline, config)
+impute_mapping_types(pipeline, config)
+impute_variant_types(pipeline, config)
 
 # add tmp dir:
 if not 'tmp_dir' in config:
